updated like_util

+++ instapy/like_util.py
@@ -29,7 +29,7 @@ def get_links_for_tag(browser, tag, amount):
 
   main_elem = browser.find_element_by_tag_name('main')
 
-  new_needed = ceil((amount - 33) / 12)
+  new_needed = int(ceil((amount - 33) / 12))
 
   for _ in range(new_needed):  # add images x * 12
     body_elem.send_keys(Keys.END)
